handling the case when trying to delete tenant without specifying the name

+++ poem/Poem/api/internal_views/tenants.py
@@ -70,24 +70,31 @@ class ListTenants(APIView):
 
         return Response(results)
 
-    def delete(self, request, name):
-        if request.tenant.schema_name == get_public_schema_name():
-            if request.user.is_superuser:
-                tenant = Tenant.objects.get(name=name)
-                tenant.delete()
+    def delete(self, request, name=None):
+        if name:
+            if request.tenant.schema_name == get_public_schema_name():
+                if request.user.is_superuser:
+                    tenant = Tenant.objects.get(name=name)
+                    tenant.delete()
 
-                return Response(status=status.HTTP_204_NO_CONTENT)
+                    return Response(status=status.HTTP_204_NO_CONTENT)
+
+                else:
+                    return error_response(
+                        status_code=status.HTTP_401_UNAUTHORIZED,
+                        detail='You do not have permission to delete tenants.'
+                    )
 
             else:
                 return error_response(
-                    status_code=status.HTTP_401_UNAUTHORIZED,
-                    detail='You do not have permission to delete tenants.'
+                    status_code=status.HTTP_403_FORBIDDEN,
+                    detail='Cannot delete tenant outside public schema.'
                 )
 
         else:
             return error_response(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail='Cannot delete tenant outside public schema.'
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail='Tenant name should be specified.'
             )
